Fix security vulnerability in react-scripts
Prior to this change, there was still a security vulnerability in merge.
React-scripts needs to be updated and browserslist added.

This change fixes the security vulnerability, updates the react-scripts,
and adds browserslist to the package.json.

https://nvd.nist.gov/vuln/detail/CVE-2018-16469

The merge.recursive function in the merge package in versions before
1.2.1 can be tricked into adding or modifying properties of the Object
prototype. These properties will be present on all objects allowing
for a denial of service attack.